fix: reuse this save function
